+++ @navikt/core/react/src/accordion/AccordionContent.tsx
@@ -1,5 +1,6 @@
 import cl from "clsx";
 import React, { forwardRef, useContext } from "react";
+import { UNSAFE_useAkselTheme } from "../provider";
 import { BodyLong } from "../typography";
 import { AccordionItemContext } from "./AccordionItem";
 
@@ -14,6 +15,8 @@ export interface AccordionContentProps
 const AccordionContent = forwardRef<HTMLDivElement, AccordionContentProps>(
   ({ children, className, ...rest }, ref) => {
     const context = useContext(AccordionItemContext);
+
+    const themeContext = UNSAFE_useAkselTheme(false);
 
     if (context === null) {
       console.error(
@@ -38,7 +41,11 @@ const AccordionContent = forwardRef<HTMLDivElement, AccordionContentProps>(
           !context.open || undefined
         } /* Added to fix bug with Radio component, where label text inside a span sometimes is ignored by screen readers after hiding/displaying the RadioGroup inside an Accordion */
       >
-        {children}
+        {themeContext ? (
+          <div className="navds-accordion__content-inner">{children}</div>
+        ) : (
+          children
+        )}
       </BodyLong>
     );
   },