+++ aksel.nav.no/website/sanity/plugins/publication-flow/actions/createPublishWithDateAction.ts
@@ -1,0 +1,23 @@
+import { DocumentActionComponent, useDocumentOperation } from "sanity";
+
+export function createPublishWithDateAction(
+  originalAction: DocumentActionComponent,
+): DocumentActionComponent {
+  return (props) => {
+    const originalResult = originalAction(props);
+
+    const { patch } = useDocumentOperation(props.id, props.type);
+    if (!originalResult) {
+      return null;
+    }
+    return {
+      ...originalResult,
+      onHandle: () => {
+        patch.execute([
+          { setIfMissing: { publishedAt: new Date().toISOString() } },
+        ]);
+        originalResult?.onHandle?.();
+      },
+    };
+  };
+}